requirements.txt
# NewsBot Intelligence System 2.0 — Dependencies
# ITAI 2373 Final Project | Houston Community College
# Install: pip install -r requirements.txt
# After install: python -m spacy download en_core_web_sm

# ── Core NLP ──────────────────────────────────────────────────────────────
spacy==3.8.10
nltk==3.8.1
vaderSentiment==3.3.2
textblob==0.18.0

# ── Machine Learning ──────────────────────────────────────────────────────
scikit-learn==1.7.2
numpy==2.4.4
scipy==1.16.1

# ── Data Processing ───────────────────────────────────────────────────────
pandas==2.3.3
pyarrow==22.0.0

# ── Topic Modeling ────────────────────────────────────────────────────────
pyLDAvis==3.4.1
gensim==0.8.8

# ── LLM Integration (Module B) ────────────────────────────────────────────
ollama==0.6.2
# Note: ollama server must be running locally (ollama serve)
# Pull model: ollama pull llama3.2

# ── Multilingual (Module C) ───────────────────────────────────────────────
#langdetect
deep-translator==1.11.4
lingua-language-detector==2.2.0

# ── Embeddings & Semantic Search ─────────────────────────────────────────
sentence-transformers==3.0.1
faiss-cpu==1.12.0

# ── Visualization ─────────────────────────────────────────────────────────
matplotlib==3.10.5
seaborn==0.13.2
plotly==5.22.0
wordcloud==1.9.5

# ── Notebook & Dev Tools ──────────────────────────────────────────────────
jupyter==1.0.0
ipywidgets==8.1.3
tqdm==4.66.4

# ── Testing ───────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-cov==5.0.0

# ── Utilities ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pyyaml==6.0.3
requests==2.32.2
contractions==0.1.73
flask==3.0.3
flask-session==0.8.0
